HYP-133: introducing fixtures to reduce code duplication.
